Hieu

+++ src/main/java/com/lowquality/serverwebm/service/MangaService.java
@@ -6,6 +6,7 @@ import com.lowquality.serverwebm.repository.ChapterRepository;
 import com.lowquality.serverwebm.repository.MangadetailRepository;
 import com.lowquality.serverwebm.util.SecurityUtils;
 import com.lowquality.serverwebm.util.UrlUtils;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.*;
 import org.springframework.data.rest.webmvc.ResourceNotFoundException;
@@ -22,31 +23,25 @@ import java.util.stream.Collectors;
 public class MangaService {
     @Autowired
     MangadetailRepository mangadetailRepository;
-    @Autowired
-    ChapterRepository chapterRepository;
-    @Autowired
-    CategoryService categoryService;
-    @Autowired
-    AuthorService authorService;
-    @Autowired
-    StatusService statusService;
-    @Autowired
-    private PermissionService permissionService;
-    @Autowired
-    private UserService userService;
-    @Autowired
-    private FileStorageService fileStorageService;
-    @Autowired
-    private UrlUtils urlUtils;
+    private final CategoryService categoryService;
+    private final AuthorService authorService;
+    private final StatusService statusService;
+    private final PermissionService permissionService;
+    private final FileStorageService fileStorageService;
+    private final ChapterService chapterService;
 
-    //    @Autowired
-//    MangaService(MangadetailRepository mangadetailRepository, ChapterRepository chapterRepository, AuthorService authorService, CategoryService categoryService, StatusService statusService) {
-//        this.mangadetailRepository = mangadetailRepository;
-//        this.chapterRepository = chapterRepository;
-//        this.categoryService = categoryService;
-//        this.authorService = authorService;
-//        this.statusService = statusService;
-//    }
+
+    @Autowired
+    MangaService(MangadetailRepository mangadetailRepository , AuthorService authorService, CategoryService categoryService, StatusService statusService, PermissionService permissionService, FileStorageService fileStorageService, ChapterService chapterService) {
+        this.mangadetailRepository = mangadetailRepository;
+
+        this.categoryService = categoryService;
+        this.authorService = authorService;
+        this.statusService = statusService;
+            this.permissionService = permissionService;
+            this.fileStorageService = fileStorageService;
+            this.chapterService = chapterService;
+        }
     public Mangadetail getMangaEntityById(Integer id) {
         return mangadetailRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Manga not found: " + id));
@@ -80,12 +75,15 @@ public class MangaService {
         mangadetailRepository.save(manga);
     }
 // cập nhật tác giả
+
     public void addAuthorToManga(Integer mangaId, String authorName) {
         Mangadetail manga = this.getMangaEntityById(mangaId);
         permissionService.checkUserPermission(manga.getUploader().getId(),"cập nhật tác giả");
         Author author = new Author();
         author.setAuthorName(authorName);
+        authorService.save(author);
         manga.setAuthor_id(author);
+
         mangadetailRepository.save(manga);
     }
 
@@ -113,6 +111,10 @@ public class MangaService {
                 .map(categoryService::convertToDTO)
                 .collect(Collectors.toList());
         String coverimgURL = UrlUtils.toPublicUrl(mangadetail.getCover_img());
+        List<ChapterDTO> chapterDTOs = new ArrayList<>();
+        if (mangadetail.getChapters() != null) {
+            chapterDTOs = chapterService.getChaptersByMangaId(mangadetail.getId());
+        }
         return MangadetailDTO.builder()
                 .id(mangadetail.getId())
                 .name(mangadetail.getName())
@@ -122,6 +124,7 @@ public class MangaService {
                 .id_category(categoryDTOs)
                 .uploader(mangadetail.getUploader().getFullName())
                 .id_status(statusDTO)
+                .chapter(chapterDTOs)
                 .build();
     }
 //    //filter
@@ -202,13 +205,14 @@ public Page<MangadetailDTO> filterManga(String search, List<Integer> categoryIds
     Pageable pageable = switch (sortBy) {
         case "latest" -> PageRequest.of(page, size, Sort.by(Sort.Direction.DESC, "updatedAt"));
         case "name"   -> PageRequest.of(page, size, Sort.by(Sort.Direction.ASC, "name"));
-        case "popular" -> PageRequest.of(page, size, Sort.by(Sort.Direction.DESC, "view") ); // chưa có view.
+//        case "popular" -> PageRequest.of(page, size, Sort.by(Sort.Direction.DESC, "view") ); // chưa có view.
         default       -> PageRequest.of(page, size);  // fallback: không sort hoặc sort custom bằng Java
     };
-
+    Long categorySize = (categoryIds == null || categoryIds.isEmpty()) ? null : (long) categoryIds.size();
     Page<Mangadetail> mangaPage = mangadetailRepository.filterMangaJPQL(
             search,
-            categoryIds != null && !categoryIds.isEmpty() ? categoryIds : null,
+            (categoryIds != null && !categoryIds.isEmpty()) ? categoryIds : null,
+            categorySize,
             statusId,
             authorId,
             uploader,
@@ -231,31 +235,18 @@ public Page<MangadetailDTO> filterManga(String search, List<Integer> categoryIds
 }
 
 public MangadetailDTO addManga(CreateMangaRequest request) {
-
         User user = SecurityUtils.getCurrentUser();
         // Tạo manga mới
         permissionService.checkMangaPermission("thêm truyện");
-        String MangaSubDir = "Manga_" + request.getName();
-        String coverImgUrl =  fileStorageService.storeFile( request.getCoverImg(),MangaSubDir);
         Mangadetail manga = new Mangadetail();
         manga.setName(request.getName());
         manga.setDescription(request.getDescription());
+         if(request.getName()!=null && request.getCoverImg()!=null) {
+        String MangaSubDir = "Manga_" + fileStorageService.sanitizeFileName(request.getName());
+        String coverImgUrl =  fileStorageService.storeFile( request.getCoverImg(),MangaSubDir);
         manga.setCover_img(coverImgUrl);
+         }
         manga.setUploader(user);
-        // Set author nếu có
-        if (request.getAuthorName() != null) {
-            Author author = authorService.findByAuthor_name(request.getAuthorName());
-            if (author != null) {
-                manga.setAuthor_id(author);
-            }
-            else{
-                AuthorDTO authorDTO=  authorService.createAuthor(request.getAuthorName());
-                author = authorService.findById(authorDTO.getId());
-                manga.setAuthor_id(author);
-            }
-
-        }
-
 
         if (request.getStatusId() != null) {
             Status status = statusService.findById(request.getStatusId());
@@ -267,23 +258,38 @@ public MangadetailDTO addManga(CreateMangaRequest request) {
         manga.setStatus_id(defaultStatus);
         // Lưu manga
         manga = mangadetailRepository.save(manga);
-
+         if (request.getAuthorName() != null && !request.getAuthorName().isBlank()) {
+        Author author = authorService.findByAuthor_name(request.getAuthorName().trim());
+        if (author == null) {
+            // Nếu chưa có, tạo mới
+            Author newAuthor = new Author();
+            newAuthor.setAuthorName(request.getAuthorName().trim());
+            author = authorService.save(newAuthor); // Lưu vào DB
+        }
+        manga.setAuthor_id(author);
+        }
         // Thêm categories nếu có
         if (request.getCategoryIds() != null && !request.getCategoryIds().isEmpty()) {
             this.addCategories(manga.getId(), request.getCategoryIds());
         }
         return convertMangadetailToDTO(manga);
     }
+
     public void editMangadetail(Integer id,CreateMangaRequest mangaRequest) {
             //CreateMangaRequest tận dụng lại để cập nhật thông tin chapter
             permissionService.checkMangaPermission("cập nhật truyện");
             Mangadetail manga = getMangaEntityById(id);
              manga.setName(mangaRequest.getName());
              manga.setDescription(mangaRequest.getDescription());
-             if(mangaRequest.getAuthorName() != null) {
-                 Author author = authorService.findByAuthor_name(mangaRequest.getAuthorName());
-                 manga.setAuthor_id(author);
-             }
+            if (mangaRequest.getAuthorName() != null && !mangaRequest.getAuthorName().isBlank()) {
+            Author author = authorService.findByAuthor_name(mangaRequest.getAuthorName().trim());
+            if (author == null) {
+                Author newAuthor = new Author();
+                newAuthor.setAuthorName(mangaRequest.getAuthorName().trim());
+                author = authorService.save(newAuthor);
+            }
+            manga.setAuthor_id(author);
+         }
              if (mangaRequest.getStatusId() != null) {
                  Status status = statusService.findById(mangaRequest.getStatusId());
                  manga.setStatus_id(status);
@@ -292,8 +298,11 @@ public MangadetailDTO addManga(CreateMangaRequest request) {
                 this.addCategories(manga.getId(), mangaRequest.getCategoryIds());
             }
              if(mangaRequest.getCoverImg() != null && !mangaRequest.getCoverImg().isEmpty()) {
-                 fileStorageService.deleteFile(manga.getCover_img());
-                 String MangaSubDir = "Manga_" + mangaRequest.getName();
+                 if (manga.getCover_img() != null) {
+                     fileStorageService.deleteFile(manga.getCover_img());
+                 }
+
+                 String MangaSubDir = "Manga_" + fileStorageService.sanitizeFileName(mangaRequest.getName());
                  String coverImgUrl =  fileStorageService.storeFile( mangaRequest.getCoverImg(),MangaSubDir);
                  manga.setCover_img(coverImgUrl);
              }
